Extract shared EmailHeader, EmailFooter, and emailStyles to eliminate duplication across email templates
