+++ lib/email/components/emailStyles.ts
@@ -1,0 +1,55 @@
+export const main = {
+  fontFamily: '"Geist", Arial, sans-serif',
+  lineHeight: 1.6,
+  color: "#09090b",
+  backgroundColor: "#fafafa",
+};
+
+export const container = {
+  maxWidth: "600px",
+  margin: "0 auto",
+  padding: "32px 24px",
+  backgroundColor: "#ffffff",
+  borderRadius: "12px",
+  boxShadow: "0 2px 4px rgba(0, 0, 0, 0.1)",
+};
+
+export const content = {
+  padding: "0 16px",
+};
+
+export const text = {
+  margin: "16px 0",
+  color: "#09090b",
+  fontSize: "16px",
+};
+
+export const details = {
+  backgroundColor: "#f9fafb",
+  borderRadius: "8px",
+  padding: "16px",
+  margin: "24px 0",
+};
+
+export const h3 = {
+  margin: "0 0 12px 0",
+  color: "#18181b",
+  fontSize: "16px",
+};
+
+export const buttonContainer = {
+  textAlign: "center" as const,
+  margin: "32px 0",
+};
+
+export const button = {
+  backgroundColor: "#3b82f6",
+  borderRadius: "6px",
+  color: "#fff",
+  fontSize: "16px",
+  textDecoration: "none",
+  textAlign: "center" as const,
+  display: "inline-block",
+  padding: "12px 24px",
+  fontWeight: 500,
+};
