+++ lib/email/templates/contactForm.tsx
@@ -10,6 +10,16 @@ import {
 } from "@react-email/components";
 
 import { ContactFormEmailProps } from "@/types";
+import { EmailFooter } from "../components/EmailFooter";
+import { EmailHeader } from "../components/EmailHeader";
+import {
+  container,
+  content,
+  details,
+  h3,
+  main,
+  text,
+} from "../components/emailStyles";
 
 export default function ContactFormEmail({
   firstName,
@@ -27,12 +37,11 @@ export default function ContactFormEmail({
       </Preview>
       <Body style={main}>
         <Container style={container}>
-          <div style={header}>
-            <div style={brand}>
-              Invoice<span style={brandSpan}>WeMaAd</span>
-            </div>
-            <div style={badge}>Contact Form</div>
-          </div>
+          <EmailHeader
+            badgeLabel="Contact Form"
+            badgeBackground="#dbeafe"
+            badgeColor="#1d4ed8"
+          />
 
           <div style={content}>
             <Heading as="h2">New Contact Form Submission</Heading>
@@ -62,87 +71,13 @@ export default function ContactFormEmail({
               </Text>
             </div>
 
-            <div style={footer}>
-              <Text style={{ ...text, color: "#71717a", fontSize: "14px" }}>
-                © {new Date().getFullYear()} InvoiceWeMaAd. All rights reserved.
-              </Text>
-              <Text style={{ ...text, color: "#71717a", fontSize: "14px" }}>
-                Making invoicing super easy!
-              </Text>
-            </div>
+            <EmailFooter />
           </div>
         </Container>
       </Body>
     </Html>
   );
 }
-
-const main = {
-  fontFamily: '"Geist", Arial, sans-serif',
-  lineHeight: 1.6,
-  color: "#09090b",
-  backgroundColor: "#fafafa",
-};
-
-const container = {
-  maxWidth: "600px",
-  margin: "0 auto",
-  padding: "32px 24px",
-  backgroundColor: "#ffffff",
-  borderRadius: "12px",
-  boxShadow: "0 2px 4px rgba(0, 0, 0, 0.1)",
-};
-
-const header = {
-  textAlign: "left" as const,
-  marginBottom: "32px",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "space-between",
-};
-
-const brand = {
-  fontSize: "24px",
-  fontWeight: 600,
-  color: "#18181b",
-};
-
-const brandSpan = {
-  color: "#3b82f6",
-};
-
-const badge = {
-  display: "inline-block",
-  padding: "6px 12px",
-  backgroundColor: "#dbeafe",
-  color: "#1d4ed8",
-  borderRadius: "9999px",
-  fontSize: "14px",
-  fontWeight: 500,
-};
-
-const content = {
-  padding: "0 16px",
-};
-
-const details = {
-  backgroundColor: "#f9fafb",
-  borderRadius: "8px",
-  padding: "16px",
-  margin: "24px 0",
-};
-
-const h3 = {
-  margin: "0 0 12px 0",
-  color: "#18181b",
-  fontSize: "16px",
-};
-
-const text = {
-  margin: "16px 0",
-  color: "#09090b",
-  fontSize: "16px",
-};
 
 const notice = {
   backgroundColor: "#eff6ff",
@@ -155,9 +90,4 @@ const notice = {
 const hr = {
   margin: "20px 0",
   borderColor: "#e4e4e7",
-};
-
-const footer = {
-  marginTop: "48px",
-  textAlign: "center" as const,
 };
