+++ lib/email/templates/reminderInvoice.tsx
@@ -10,6 +10,18 @@ import {
 } from "@react-email/components";
 
 import { ReminderInvoiceEmailProps } from "@/types";
+import { EmailFooter } from "../components/EmailFooter";
+import { EmailHeader } from "../components/EmailHeader";
+import {
+  button,
+  buttonContainer,
+  container,
+  content,
+  details,
+  h3,
+  main,
+  text,
+} from "../components/emailStyles";
 
 export default function ReminderInvoiceEmail({
   clientName,
@@ -26,12 +38,11 @@ export default function ReminderInvoiceEmail({
       <Preview>Payment Reminder for Invoice #{invoiceNumber}</Preview>
       <Body style={main}>
         <Container style={container}>
-          <div style={header}>
-            <div style={brand}>
-              Invoice<span style={brandSpan}>WeMaAd</span>
-            </div>
-            <div style={urgentBadge}>Payment Reminder</div>
-          </div>
+          <EmailHeader
+            badgeLabel="Payment Reminder"
+            badgeBackground="#ffe4e6"
+            badgeColor="#be123c"
+          />
 
           <div style={content}>
             <Heading as="h2">Invoice Payment Reminder</Heading>
@@ -71,14 +82,7 @@ export default function ReminderInvoiceEmail({
               </Link>
             </div>
 
-            <div style={footer}>
-              <Text style={{ ...text, color: "#71717a", fontSize: "14px" }}>
-                © {new Date().getFullYear()} InvoiceWeMaAd. All rights reserved.
-              </Text>
-              <Text style={{ ...text, color: "#71717a", fontSize: "14px" }}>
-                Making invoicing super easy!
-              </Text>
-            </div>
+            <EmailFooter />
           </div>
         </Container>
       </Body>
@@ -86,99 +90,10 @@ export default function ReminderInvoiceEmail({
   );
 }
 
-const main = {
-  fontFamily: '"Geist", Arial, sans-serif',
-  lineHeight: 1.6,
-  color: "#09090b",
-  backgroundColor: "#fafafa",
-};
-
-const container = {
-  maxWidth: "600px",
-  margin: "0 auto",
-  padding: "32px 24px",
-  backgroundColor: "#ffffff",
-  borderRadius: "12px",
-  boxShadow: "0 2px 4px rgba(0, 0, 0, 0.1)",
-};
-
-const header = {
-  textAlign: "left" as const,
-  marginBottom: "32px",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "space-between",
-};
-
-const brand = {
-  fontSize: "24px",
-  fontWeight: 600,
-  color: "#18181b",
-};
-
-const brandSpan = {
-  color: "#3b82f6",
-};
-
-const urgentBadge = {
-  display: "inline-block",
-  padding: "6px 12px",
-  backgroundColor: "#ffe4e6",
-  color: "#be123c",
-  borderRadius: "9999px",
-  fontSize: "14px",
-  fontWeight: 500,
-};
-
-const content = {
-  padding: "0 16px",
-};
-
-const details = {
-  backgroundColor: "#f9fafb",
-  borderRadius: "8px",
-  padding: "16px",
-  margin: "24px 0",
-};
-
-const h3 = {
-  margin: "0 0 12px 0",
-  color: "#18181b",
-  fontSize: "16px",
-};
-
-const text = {
-  margin: "16px 0",
-  color: "#09090b",
-  fontSize: "16px",
-};
-
 const urgentNotice = {
   backgroundColor: "#ffe4e6",
   borderLeft: "4px solid #be123c",
   padding: "16px",
   margin: "24px 0",
   borderRadius: "0 8px 8px 0",
-};
-
-const buttonContainer = {
-  textAlign: "center" as const,
-  margin: "32px 0",
-};
-
-const button = {
-  backgroundColor: "#3b82f6",
-  borderRadius: "6px",
-  color: "#fff",
-  fontSize: "16px",
-  textDecoration: "none",
-  textAlign: "center" as const,
-  display: "inline-block",
-  padding: "12px 24px",
-  fontWeight: 500,
-};
-
-const footer = {
-  marginTop: "48px",
-  textAlign: "center" as const,
 };
