+++ lib/email/templates/updatedInvoice.tsx
@@ -10,6 +10,19 @@ import {
 } from "@react-email/components";
 
 import { UpdatedInvoiceEmailProps } from "@/types";
+import { EmailFooter } from "../components/EmailFooter";
+import { EmailHeader } from "../components/EmailHeader";
+import {
+  button,
+  buttonContainer,
+  container,
+  content,
+  details,
+  h3,
+  main,
+  text,
+} from "../components/emailStyles";
+
 export default function UpdatedInvoiceEmail({
   clientName,
   invoiceNumber,
@@ -25,12 +38,11 @@ export default function UpdatedInvoiceEmail({
       <Preview>Invoice #{invoiceNumber} has been updated</Preview>
       <Body style={main}>
         <Container style={container}>
-          <div style={header}>
-            <div style={brand}>
-              Invoice<span style={brandSpan}>WeMaAd</span>
-            </div>
-            <div style={warningBadge}>Updated Invoice</div>
-          </div>
+          <EmailHeader
+            badgeLabel="Updated Invoice"
+            badgeBackground="#fef3c7"
+            badgeColor="#b45309"
+          />
 
           <div style={content}>
             <Heading as="h2">Invoice Updated</Heading>
@@ -69,14 +81,7 @@ export default function UpdatedInvoiceEmail({
               </Link>
             </div>
 
-            <div style={footer}>
-              <Text style={{ ...text, color: "#71717a", fontSize: "14px" }}>
-                © {new Date().getFullYear()} InvoiceWeMaAd. All rights reserved.
-              </Text>
-              <Text style={{ ...text, color: "#71717a", fontSize: "14px" }}>
-                Making invoicing super easy!
-              </Text>
-            </div>
+            <EmailFooter />
           </div>
         </Container>
       </Body>
@@ -84,99 +89,10 @@ export default function UpdatedInvoiceEmail({
   );
 }
 
-const main = {
-  fontFamily: '"Geist", Arial, sans-serif',
-  lineHeight: 1.6,
-  color: "#09090b",
-  backgroundColor: "#fafafa",
-};
-
-const container = {
-  maxWidth: "600px",
-  margin: "0 auto",
-  padding: "32px 24px",
-  backgroundColor: "#ffffff",
-  borderRadius: "12px",
-  boxShadow: "0 2px 4px rgba(0, 0, 0, 0.1)",
-};
-
-const header = {
-  textAlign: "left" as const,
-  marginBottom: "32px",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "space-between",
-};
-
-const brand = {
-  fontSize: "24px",
-  fontWeight: 600,
-  color: "#18181b",
-};
-
-const brandSpan = {
-  color: "#3b82f6",
-};
-
-const warningBadge = {
-  display: "inline-block",
-  padding: "6px 12px",
-  backgroundColor: "#fef3c7",
-  color: "#b45309",
-  borderRadius: "9999px",
-  fontSize: "14px",
-  fontWeight: 500,
-};
-
-const content = {
-  padding: "0 16px",
-};
-
-const details = {
-  backgroundColor: "#f9fafb",
-  borderRadius: "8px",
-  padding: "16px",
-  margin: "24px 0",
-};
-
-const h3 = {
-  margin: "0 0 12px 0",
-  color: "#18181b",
-  fontSize: "16px",
-};
-
-const text = {
-  margin: "16px 0",
-  color: "#09090b",
-  fontSize: "16px",
-};
-
 const warningNotice = {
   backgroundColor: "#fef3c7",
   borderLeft: "4px solid #b45309",
   padding: "16px",
   margin: "24px 0",
   borderRadius: "0 8px 8px 0",
-};
-
-const buttonContainer = {
-  textAlign: "center" as const,
-  margin: "32px 0",
-};
-
-const button = {
-  backgroundColor: "#3b82f6",
-  borderRadius: "6px",
-  color: "#fff",
-  fontSize: "16px",
-  textDecoration: "none",
-  textAlign: "center" as const,
-  display: "inline-block",
-  padding: "12px 24px",
-  fontWeight: 500,
-};
-
-const footer = {
-  marginTop: "48px",
-  textAlign: "center" as const,
 };
